# WebServer: Allow client to send many requests on the same connection