--- a/libraries/ESP8266WebServer/src/ESP8266WebServer-impl.h
+++ b/libraries/ESP8266WebServer/src/ESP8266WebServer-impl.h
@@ -49,6 +49,7 @@ ESP8266WebServerTemplate<ServerType>::ESP8266WebServerTemplate(IPAddress addr, i
 , _currentVersion(0)
 , _currentStatus(HC_NONE)
 , _statusChange(0)
+, _keepAlive(false)
 , _currentHandler(nullptr)
 , _firstHandler(nullptr)
 , _lastHandler(nullptr)
@@ -325,6 +326,10 @@ void ESP8266WebServerTemplate<ServerType>::handleClient() {
   bool callYield = false;
 
   if (_currentClient.connected() || _currentClient.available()) {
+    if (_currentClient.available() && _keepAlive) {
+      _currentStatus = HC_WAIT_READ;
+    }
+
     switch (_currentStatus) {
     case HC_NONE:
       // No-op to avoid C++ compiler warning
@@ -362,6 +367,7 @@ void ESP8266WebServerTemplate<ServerType>::handleClient() {
   if (!keepCurrentClient) {
     _currentClient = ClientType();
     _currentStatus = HC_NONE;
+    _keepAlive = false;
     _currentUpload.reset();
   }
 
@@ -374,6 +380,7 @@ template <typename ServerType>
 void ESP8266WebServerTemplate<ServerType>::close() {
   _server.close();
   _currentStatus = HC_NONE;
+  _keepAlive = false;
   if(!_headerKeysCount)
     collectHeaders(0, 0);
 }
@@ -429,7 +436,14 @@ void ESP8266WebServerTemplate<ServerType>::_prepareHeader(String& response, int 
     if (_corsEnabled) {
       sendHeader(String(F("Access-Control-Allow-Origin")), String("*"));
     }
-    sendHeader(String(F("Connection")), String(F("close")));
+
+    if (_keepAlive && _server.hasClient()) { // Disable keep alive if another client is waiting.
+      _keepAlive = false;
+    }
+    sendHeader(String(F("Connection")), String(_keepAlive ? F("keep-alive") : F("close")));
+    if (_keepAlive) {
+      sendHeader(String(F("Keep-Alive")), String(F("timeout=")) + HTTP_MAX_CLOSE_WAIT);
+    }
 
     response += _responseHeaders;
     response += "\r\n";