--- a/libraries/ESP8266WebServer/src/ESP8266WebServer.h
+++ b/libraries/ESP8266WebServer/src/ESP8266WebServer.h
@@ -170,8 +170,9 @@ public:
 
   // Whether other requests should be accepted from the client on the
   // same socket after a response is sent.
-  // This will automatically configure the "Connection" header.
-  // Defaults to false.
+  // This will automatically configure the "Connection" header of the response.
+  // Defaults to true when the client's HTTP version is 1.1 or above, otherwise it defaults to false.
+  // If the client sends the "Connection" header, the value given by the header is used.
   void keepAlive(bool keepAlive) { _keepAlive = keepAlive; }
   bool keepAlive() { return _keepAlive; }