@d-a-v 2020-06-29
Thanks!

In `::_prepareHeader()` should `keepalive` be assumed to be false when http version is 1.0 ?
I wonder why you reset it to false at every connection, I would think it is safe to leave it to true once for all.
It is false by default, that's OK if we merge this for core-v2.7.2, but I would set it to true by default in core-v3.