@ZakCodes 2020-06-29
Just like the headers for a response are reset for each connection, I thought that it would be better to decide whether to keep a connection alive for each connection. Also, it's important to set it to `false` if there's another incoming connection, so we can't let the user assume that they can set `keepAlive` to `true` once when the server is being setup and that it will be true forever.

However, it would be a good idea to set it to `true` by default in `core-v3.0.0` just like you said. Do you want me to do another pull request for that?