@d-a-v 2020-07-10
>  I thought that it would be better to decide whether to keep a connection alive for each connection.

That's true but I don't understand what prevents it to be true by default.
You are always checking for a new client, then you disable keep-alive when there's one.
Isn't that enough ?

> However, it would be a good idea to set it to true by default in core-v3.0.0 just like you said. Do you want me to do another pull request for that?

You can update this PR with the new default value. 2.7.2 is released and we are now in 3.0.0-dev.