@ZakCodes 2020-07-10
I've made changes to keep the connection alive by default. Does this look good to you?