@d-a-v 2020-07-10
> Does this look good to you?

The "Default to false" comment needs change.

I wonder if this test
```
    if (_keepAlive && _server.hasClient()) { // Disable keep alive if another client is waiting.
      _keepAlive = false;
    }
```
should also check `_currentVersion` (0 for http/1.0, 1 for http/1.1)
```
    if (_currentVersion == 0 || _server.hasClient()) { // Disable keep alive if HTTP/1.0 or another client is waiting.
      _keepAlive = false;
    }
```

Apart from that it looks good !