@ZakCodes 2020-07-10
I get what you mean.

I changed the place where `_keepAlive`'s value is initialized for each request. When the request is parsed, `_keepAlive` will be set to `true` if the HTTP version is above 1.0 and it will be set to `false` otherwise. Then, the received headers will be checked to see whether the `Connection` header has been received from the client. If the client has sent `Connection: keep-alive`, `_keepAlive` is set to true, if `Connection: {anything other than keep-alive}` is received, `_keepAlive` is set to false.