@ZakCodes 2020-07-16
As mentioned in my last comment, the keep-alive feature that I've implemented is now automatically activated if the HTTP 1.1 or above or if the `Connection: Keep-Alive` header is in the request, so everyone will see the benefits of this PR when they update to the new version of this library without having to do anything else.

However, we could show an example of disabling it. Disabling it is useful when a request couldn't be authenticated. By disabling it, you force the attacker to make another full TLS handshake to try another password which would then require more time for the attacker to crack the password.  

Therefore, we could disable keep-alive when the authentication is failed in the [SimpleAuthentication](https://github.com/esp8266/Arduino/blob/master/libraries/ESP8266WebServer/examples/SimpleAuthentication/SimpleAuthentication.ino) example.  
We could also disable keep-alive in [`ESP8266WebServerTemplate::requestAuthentication`](https://github.com/esp8266/Arduino/blob/master/libraries/ESP8266WebServer/src/ESP8266WebServer-impl.h#L255).