@earlephilhower 2020-07-16
Thanks for the clarification, @ZakCodes .  I don't think we need to show disabling because in general it's not a good thing to do performance-wise and there's not really a downside I can see now.