@TD-er 2020-07-16
"Brute force attack" on an ESP node also sounds like a good way to get its service down.
Especially with TLS enabled, you don't need much 'brute' force.