@ZakCodes 2020-07-16
> @ZakCodes Passwords should not be that simple that they can be cracked by a brute-force attack. Then you should use better passwords instead.

Good point